requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.12.0
boto3==1.42.17
botocore==1.42.17
click==8.3.1
fastapi==0.128.0
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
opencv-python==4.10.0.84
pillow==11.2.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.5.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
mkdocs==1.5.3
mkdocs-material==9.5.3
